adding unsafe ba version
