+++ enzyme/benchmarks/ReverseMode/adbench/ba.h
@@ -127,19 +127,16 @@ extern "C" {
         double* reproj_err,
         double* w_err
     );
-    
-    void rust2_ba_objective(
-        int n,
-        int m,
-        int p,
-        double const* cams,
-        double const* X,
-        double const* w,
-        int const* obs,
-        double const* feats,
-        double* reproj_err,
-        double* w_err
-    );
+
+    void rust2_unsafe_ba_objective(int n, int m, int p, double const *cams,
+                                   double const *X, double const *w,
+                                   int const *obs, double const *feats,
+                                   double *reproj_err, double *w_err);
+
+    void rust2_ba_objective(int n, int m, int p, double const *cams,
+                            double const *X, double const *w, int const *obs,
+                            double const *feats, double *reproj_err,
+                            double *w_err);
 
     void dcompute_reproj_error(
         double const* cam,
@@ -183,17 +180,17 @@ extern "C" {
 
     void adept_compute_zach_weight_error(double const* w, double* dw, double* err, double* derr);
 
-    void rust_dcompute_reproj_error(
-        double const* cam,
-        double * dcam,
-        double const* X,
-        double * dX,
-        double const* w,
-        double * wb,
-        double const* feat,
-        double *err,
-        double *derr
-    );
+    void rust_unsafe_dcompute_reproj_error(double const *cam, double *dcam,
+                                           double const *X, double *dX,
+                                           double const *w, double *wb,
+                                           double const *feat, double *err,
+                                           double *derr);
+
+    void rust_dcompute_reproj_error(double const *cam, double *dcam,
+                                    double const *X, double *dX,
+                                    double const *w, double *wb,
+                                    double const *feat, double *err,
+                                    double *derr);
 
     void rust_dcompute_zach_weight_error(double const* w, double* dw, double* err, double* derr);
 }
@@ -362,10 +359,22 @@ int main(const int argc, const char* argv[]) {
     std::string path = "/mnt/Data/git/Enzyme/apps/ADBench/data/ba/ba1_n49_m7776_p31843.txt";
 
     std::vector<std::string> paths = {
-        "ba10_n1197_m126327_p563734.txt",  "ba14_n356_m226730_p1255268.txt",   "ba18_n1936_m649673_p5213733.txt",    "ba2_n21_m11315_p36455.txt",    "ba6_n539_m65220_p277273.txt",  "test.txt",
-        "ba11_n1723_m156502_p678718.txt",  "ba15_n1102_m780462_p4052340.txt",  "ba19_n4585_m1324582_p9125125.txt",   "ba3_n161_m48126_p182072.txt",  "ba7_n93_m61203_p287451.txt",
-        "ba12_n253_m163691_p899155.txt",   "ba16_n1544_m942409_p4750193.txt",  "ba1_n49_m7776_p31843.txt",           "ba4_n372_m47423_p204472.txt",  "ba8_n88_m64298_p383937.txt",
-        "ba13_n245_m198739_p1091386.txt",  "ba17_n1778_m993923_p5001946.txt",  "ba20_n13682_m4456117_p2987644.txt",  "ba5_n257_m65132_p225911.txt",  "ba9_n810_m88814_p393775.txt",
+        "ba10_n1197_m126327_p563734.txt",
+        "ba14_n356_m226730_p1255268.txt", //  "ba18_n1936_m649673_p5213733.txt",
+                                          //  "ba2_n21_m11315_p36455.txt",
+                                          //  "ba6_n539_m65220_p277273.txt",
+                                          //  "test.txt",
+        //       "ba11_n1723_m156502_p678718.txt",
+        //       "ba15_n1102_m780462_p4052340.txt",
+        //       "ba19_n4585_m1324582_p9125125.txt",
+        //       "ba3_n161_m48126_p182072.txt",  "ba7_n93_m61203_p287451.txt",
+        //       "ba12_n253_m163691_p899155.txt",
+        //       "ba16_n1544_m942409_p4750193.txt",  "ba1_n49_m7776_p31843.txt",
+        //       "ba4_n372_m47423_p204472.txt",  "ba8_n88_m64298_p383937.txt",
+        //       "ba13_n245_m198739_p1091386.txt",
+        //       "ba17_n1778_m993923_p5001946.txt",
+        //       "ba20_n13682_m4456117_p2987644.txt",
+        //       "ba5_n257_m65132_p225911.txt",  "ba9_n810_m88814_p393775.txt",
     };
 
     std::ofstream jsonfile("results.json", std::ofstream::trunc);
@@ -571,7 +580,40 @@ int main(const int argc, const char* argv[]) {
     }
     }
 
-    
+    {
+    struct BAInput input;
+    read_ba_instance("data/" + path, input.n, input.m, input.p, input.cams,
+                     input.X, input.w, input.obs, input.feats);
+
+    struct BAOutput result = {std::vector<double>(2 * input.p),
+                              std::vector<double>(input.p),
+                              BASparseMat(input.n, input.m, input.p)};
+    {
+
+      struct timeval start, end;
+      gettimeofday(&start, NULL);
+      rust2_unsafe_ba_objective(input.n, input.m, input.p, input.cams.data(),
+                                input.X.data(), input.w.data(),
+                                input.obs.data(), input.feats.data(),
+                                result.reproj_err.data(), result.w_err.data());
+      gettimeofday(&end, NULL);
+      printf("primal unsafe rust t=%0.6f\n", tdiff(&start, &end));
+      json enzyme;
+      enzyme["name"] = "primal unsafe rust";
+      enzyme["runtime"] = tdiff(&start, &end);
+      for (unsigned i = 0; i < 5; i++) {
+        printf("%f ", result.reproj_err[i]);
+        enzyme["result"].push_back(result.reproj_err[i]);
+      }
+      for (unsigned i = 0; i < 5; i++) {
+        printf("%f ", result.w_err[i]);
+        enzyme["result"].push_back(result.w_err[i]);
+      }
+      printf("\n");
+      test_suite["tools"].push_back(enzyme);
+    }
+    }
+
     {
     struct BAInput input;
     read_ba_instance("data/" + path, input.n, input.m, input.p, input.cams, input.X, input.w, input.obs, input.feats);
@@ -629,6 +671,35 @@ int main(const int argc, const char* argv[]) {
     {
       struct timeval start, end;
       gettimeofday(&start, NULL);
+      calculate_jacobian<rust_unsafe_dcompute_reproj_error,
+                         rust_dcompute_zach_weight_error>(input, result);
+      gettimeofday(&end, NULL);
+      printf("Enzyme unsafe rust combined %0.6f\n", tdiff(&start, &end));
+      json enzyme;
+      enzyme["name"] = "Enzyme unsafe rust combined";
+      enzyme["runtime"] = tdiff(&start, &end);
+      for (unsigned i = 0; i < 5; i++) {
+        printf("%f ", result.J.vals[i]);
+        enzyme["result"].push_back(result.J.vals[i]);
+      }
+      printf("\n");
+      test_suite["tools"].push_back(enzyme);
+    }
+    }
+
+    {
+
+    struct BAInput input;
+    read_ba_instance("data/" + path, input.n, input.m, input.p, input.cams,
+                     input.X, input.w, input.obs, input.feats);
+
+    struct BAOutput result = {std::vector<double>(2 * input.p),
+                              std::vector<double>(input.p),
+                              BASparseMat(input.n, input.m, input.p)};
+
+    {
+      struct timeval start, end;
+      gettimeofday(&start, NULL);
       calculate_jacobian<rust_dcompute_reproj_error, rust_dcompute_zach_weight_error>(input, result);
       gettimeofday(&end, NULL);
       printf("Enzyme rust combined %0.6f\n", tdiff(&start, &end));
@@ -642,7 +713,6 @@ int main(const int argc, const char* argv[]) {
       printf("\n");
       test_suite["tools"].push_back(enzyme);
     }
-
     }
 
     test_suite["llvm-version"] = __clang_version__;
